+++ onlyoffice-nuxeo-core/src/main/java/org/onlyoffice/operation/ConvertOperation.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,25 +19,30 @@
 package org.onlyoffice.operation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.onlyoffice.client.DocumentServerClient;
 import com.onlyoffice.manager.document.DocumentManager;
 import com.onlyoffice.model.convertservice.ConvertRequest;
 import com.onlyoffice.model.convertservice.ConvertResponse;
 import com.onlyoffice.service.convert.ConvertService;
-import org.apache.hc.core5.http.HttpEntity;
 import org.nuxeo.ecm.automation.OperationException;
 import org.nuxeo.ecm.automation.core.Constants;
 import org.nuxeo.ecm.automation.core.annotations.Context;
 import org.nuxeo.ecm.automation.core.annotations.Operation;
 import org.nuxeo.ecm.automation.core.annotations.OperationMethod;
 import org.nuxeo.ecm.automation.core.annotations.Param;
-import org.nuxeo.ecm.core.api.*;
+import org.nuxeo.ecm.core.api.Blob;
+import org.nuxeo.ecm.core.api.Blobs;
+import org.nuxeo.ecm.core.api.CoreSession;
+import org.nuxeo.ecm.core.api.DocumentModel;
+import org.nuxeo.ecm.core.api.DocumentSecurityException;
+import org.nuxeo.ecm.core.api.IdRef;
 import org.nuxeo.ecm.core.api.blobholder.BlobHolder;
 import org.nuxeo.ecm.core.api.security.SecurityConstants;
 import org.nuxeo.runtime.api.Framework;
-import com.onlyoffice.manager.request.RequestManager;
 import org.onlyoffice.utils.Utils;
 
-import java.io.IOException;
+import java.io.File;
+import java.io.FileOutputStream;
 import java.io.Serializable;
 import java.util.Locale;
 
@@ -62,7 +67,7 @@ public class ConvertOperation {
     public String run() throws Exception {
         ConvertService convertService = Framework.getService(ConvertService.class);
         DocumentManager documentManager = Framework.getService(DocumentManager.class);
-        RequestManager requestManager = Framework.getService(RequestManager.class);
+        DocumentServerClient documentServerClient = Framework.getService(DocumentServerClient.class);
         Utils utils = Framework.getService(Utils.class);
 
         DocumentModel model = session.getDocument(new IdRef(id));
@@ -96,20 +101,23 @@ public class ConvertOperation {
         ConvertResponse convertResponse = convertService.processConvert(convertRequest, id);
 
         if (convertResponse.getEndConvert() != null && convertResponse.getEndConvert()) {
-            requestManager.executeGetRequest(convertResponse.getFileUrl(), new RequestManager.Callback<Void>() {
-                public Void doWork(Object response) throws IOException {
-                    Blob blob = Blobs.createBlob(((HttpEntity)response).getContent());
-                    blob.setFilename(title + "." + targetExtension);
-                    blob.setMimeType(utils.getMimeType(title + "." + targetExtension));
-
-                    model.setPropertyValue("file:content", (Serializable) blob);
-
-                    session.saveDocument(model);
-                    session.save();
-
-                    return null;
+            File tempFile = File.createTempFile("onlyoffice", null);
+            try {
+                try (FileOutputStream fos = new FileOutputStream(tempFile)) {
+                    documentServerClient.getFile(convertResponse.getFileUrl(), fos);
                 }
-            });
+
+                Blob blob = Blobs.createBlob(tempFile);
+                blob.setFilename(title + "." + targetExtension);
+                blob.setMimeType(utils.getMimeType(title + "." + targetExtension));
+
+                model.setPropertyValue("file:content", (Serializable) blob);
+
+                session.saveDocument(model);
+                session.save();
+            } finally {
+                tempFile.delete();
+            }
         }
 
         ObjectMapper objectMapper = new ObjectMapper();