+++ onlyoffice-nuxeo-core/src/main/java/org/onlyoffice/operation/CreateOperation.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -78,7 +78,7 @@ public class CreateOperation {
         DocumentType documentType = DocumentType.valueOf(type.toUpperCase());
         String extension = documentManager.getDefaultExtension(documentType);
 
-        try (InputStream inputStream = documentManager.getNewBlankFile(extension, locale)){
+        try (InputStream inputStream = documentManager.getNewBlankFile(extension, locale)) {
             DocumentModel newDoc = session.createDocumentModel(path, title, "File");
 
             Blob blob = Blobs.createBlob(inputStream);