+++ onlyoffice-nuxeo-core/src/main/java/org/onlyoffice/sdk/client/ApacheHttpclientDocumentServerClientImpl.java
@@ -16,12 +16,18 @@
  *
  */
 
-package org.onlyoffice.sdk.service.settings;
+package org.onlyoffice.sdk.client;
 
-import com.onlyoffice.model.settings.validation.ValidationResult;
+import com.onlyoffice.client.ApacheHttpclientDocumentServerClient;
+import com.onlyoffice.manager.settings.SettingsManager;
+import com.onlyoffice.manager.url.UrlManager;
+import org.nuxeo.runtime.api.Framework;
 
-import java.util.Map;
-
-public interface SettingsValidationService extends com.onlyoffice.service.settings.SettingsValidationService {
-    Map<String, ValidationResult> validateSettings();
+public class ApacheHttpclientDocumentServerClientImpl extends ApacheHttpclientDocumentServerClient {
+    public ApacheHttpclientDocumentServerClientImpl() {
+        super(
+                Framework.getService(SettingsManager.class),
+                Framework.getService(UrlManager.class)
+        );
+    }
 }