+++ onlyoffice-nuxeo-core/src/main/java/org/onlyoffice/sdk/manager/url/UrlManagerImpl.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -31,7 +31,7 @@ import org.nuxeo.ecm.webengine.WebEngine;
 import org.nuxeo.ecm.webengine.model.WebContext;
 import org.nuxeo.runtime.api.Framework;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 public class UrlManagerImpl extends DefaultUrlManager {
     private static final String APPLICATION_NAME = "onlyoffice-nuxeo";
@@ -86,7 +86,7 @@ public class UrlManagerImpl extends DefaultUrlManager {
     }
 
     @Override
-    public String getTestConvertUrl(String productInnerUrl) {
+    public String getTestConvertUrl(final String productInnerUrl) {
         WebContext ctx = WebEngine.getActiveContext();
 
         return String.format(
@@ -96,7 +96,7 @@ public class UrlManagerImpl extends DefaultUrlManager {
         );
     }
 
-    private String getBaseNuxeoUrl(WebContext ctx, Boolean inner) {
+    private String getBaseNuxeoUrl(final WebContext ctx, final Boolean inner) {
         HttpServletRequest request = ctx.getRequest();
         String webAppName = VirtualHostHelper.getWebAppName(request);
         String serverUrl = null;
@@ -111,7 +111,7 @@ public class UrlManagerImpl extends DefaultUrlManager {
         return StringUtils.isNotBlank(webAppName) ? sanitizeUrl(serverUrl) + "/" + webAppName : sanitizeUrl(serverUrl);
     }
 
-    private String generateToken(String userName) {
+    private String generateToken(final String userName) {
         TokenAuthenticationService tokenAuthenticationService = Framework.getService(TokenAuthenticationService.class);
 
         return  tokenAuthenticationService.acquireToken(